+++ vitest.config.ts
@@ -5,5 +5,8 @@ export default defineConfig({
   resolve: {
     alias: { '#': fileURLToPath(new URL('./src', import.meta.url)) }
   },
-  test: { include: ['src/**/*.test.ts'] }
+  test: {
+    include: ['src/**/*.test.ts'],
+    setupFiles: ['./src/test/vitest.setup.ts']
+  }
 });